LayoutTests/fast/page-color-sampling/color-sampling-modal-popup-expected.txt

PASS edgeColorsBeforeShowingPopup.top is "rgb(255, 0, 0)"
PASS edgeColorsBeforeShowingPopup.left is "rgb(0, 128, 0)"
PASS edgeColorsBeforeShowingPopup.right is "rgb(0, 0, 255)"
PASS edgeColorsBeforeShowingPopup.bottom is null
PASS edgeColorsAfterShowingPopup.top is "rgb(255, 0, 0)"
PASS edgeColorsAfterShowingPopup.left is "rgb(0, 128, 0)"
PASS edgeColorsAfterShowingPopup.right is "rgb(0, 0, 255)"
PASS edgeColorsAfterShowingPopup.bottom is "rgb(128, 128, 128)"
PASS edgeColorsAfterHidingPopup.top is "rgb(255, 0, 0)"
PASS edgeColorsAfterHidingPopup.left is "rgb(0, 128, 0)"
PASS edgeColorsAfterHidingPopup.right is "rgb(0, 0, 255)"
PASS edgeColorsAfterHidingPopup.bottom is null
PASS successfullyParsed is true

TEST COMPLETE
